flake: rename targetSystems to systems
The name `targetSystems` was probably chosen because of rustc's concept
of a target triple, but in this case, the concept does not map
perfectly. Most flakes out there simply use the name `systems`.

This can be revisited in the future if Nixpkgs `crossSystem` is ever
used (for cross-compiling tests as an example), but in that event, there
will be more drastic changes to support such a scheme.

Note that since this flake only provides a development shell, I am
opting to not use <https://github.com/nix-systems>, if the flake ever
provides binary packages, definitely use that pattern (so that systems
can be overridden by consumers).