facet-core: Shape: Debug: omit inner if None
If `inner` is `None` it is not shown, but if it is `Some` it is not
formatted in a matter that indicates this is an `Option`.

Is this bad practice? Should it always show `Some`?

The rational is to save horizontal characters, and the alternate display
mode is provided for accuracy; by default prefer brevity.