facet-core: Shape: Debug: show line count for doc
This shows the `doc` field as `«N lines»`, instead of taking up a bunch
of space to format the slice. I think it's safe to hide this, and if
needed, the user can `{:?}` the field directly.

One unresolved question: should this be shown as "N lines" or as
"N strings"? I ask because a `#[doc]` attribute can contain newlines,
though they are escaped. I have chosen "lines" because I saw that term
used in the documentation for the `doc` field, and I wanted to remain
consistent.