feat: add support for decoding floats in msgpack
There's the potential for nuance based on how floats are cast. I figured
it was fine that f32 gets cast up to f64 when parsing, then back down to
f32 when being placed into the partial. By IEEE 754 specification, there
should be no issues - it's just moving bits around; the bits themselves
aren't truncated in any way. The odd case is when a type is f32, but it
is decoded from an f64. In that case, edge cases like
`f32::MAX != (f64::MAX as u32)` may occur, which may be reflected in
msgpack structures generated by systems that only support
double-precision floats. In that case, there's two (acceptable) options:
throw a type error, or parse a valid f32 and ensure the value being cast
down to f32 and up to f64 does not change the value, in which case, the
f32 variant may be safely returned. Two tests have been added to reflect
this behavior: successfully parsing a f32, and a FloatOverflow when a
f64 can't be represented as an f32.